refactor: rename AppSidebar to AppNavigation
